Don't encode component locations that are same as the master's
This is both a bug fix and a significant optimization.

Is a bug fix because previously we were not encoding any
location component that was zero, even if the master had a
non-zero value for that axis.  That was wrong.
Ideally we should do that for all public-facing axes.
But the current heuristic is good enough.